Refactor useCopyToClipboard into its own hook
